Making app respond to environment
Forcing the app to find the appropriate place to store the file, depending on if it is running on Azure or locally (using an environment variable on azure to ensure files are stored outside the bin directory. Splitting the process of checking if a folder exists, and if not, then making one, out into a seperate module, so that it can be used across multiple routes.